抽取global css组件
