+++ src/App.tsx
@@ -1,27 +1,16 @@
-import React, { useMemo } from "react"
+import React from "react"
 import { useRoutes } from "react-router-dom"
-import { Global, css } from "@emotion/react"
-import { GlobalStyle } from "./theme/resetCss"
 import routers from "./router/routers"
 import Loading from "./component/loading"
 import Toast from "./component/toast"
-import { useStores } from "./models/root-store/root-store"
-import { observer } from "mobx-react-lite"
+import GlobalCssComp from "./component/globalCss"
 
-const App: React.FunctionComponent<{}> = observer(() => {
+const App: React.FunctionComponent<{}> = () => {
   const routersElement = useRoutes(routers)
 
-  const { userInfoStore } = useStores()
-  const globalStyle = useMemo(() => {
-    return GlobalStyle(userInfoStore.languageCode)
-  }, [userInfoStore.languageCode])
   return (
     <>
-      <Global
-        styles={css`
-          ${globalStyle}
-        `}
-      />
+      <GlobalCssComp />
       <Loading
         ref={(loaidngRef: ILoading) => {
           global.$loading = loaidngRef
@@ -35,6 +24,6 @@ const App: React.FunctionComponent<{}> = observer(() => {
       {routersElement}
     </>
   )
-})
+}
 
 export default App
